requirements.txt
absl-py==0.7.0
astor==0.7.1
atari-py==0.1.7
baselines==0.1.5
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.7.0
cycler==0.10.0
Cython==0.29.3
dataclasses==0.6
dill==0.2.9
fire==0.1.3
future==0.17.1
gast==0.2.2
glfw==1.7.0
grpcio==1.18.0
gym==0.10.9
h5py==2.9.0
horovod==0.15.2
idna==2.8
imageio==2.4.1
joblib==0.13.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.6
kiwisolver==1.0.1
lockfile==0.12.2
loky==2.3.1
Markdown==3.0.1
matplotlib==3.0.2
mpi4py==3.0.0
numpy==1.16.0
opencv-python==4.0.0.21
pandas==0.24.0
Pillow==5.4.1
progressbar2==3.39.2
protobuf==3.6.1
psutil==5.5.0
pycparser==2.19
pyglet==1.3.2
PyOpenGL==3.1.0
pyparsing==2.3.1
python-dateutil==2.7.5
python-utils==2.3.0
pytz==2018.9
pyzmq==17.1.2
requests==2.21.0
scipy==1.2.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
tqdm==4.30.0
urllib3==1.24.1
Werkzeug==0.14.1
zmq==0.0.0
